working: create page allocation optimization

+++ src/conversion.cpp
@@ -6,8 +6,8 @@
 // and return the bytes directly. the bytes will be copied
 conversion::comparable_key conversion::convert(const char* v, size_t vlen, bool noint)
 {
-	int64_t i;
-	double d;
+	int32_t i;
+	float d;
 
 	if (!noint && is_integer(v, vlen))
 	{
